+++ src/Service/FileProcessor.php
@@ -11,11 +11,6 @@ namespace App\Service;
 class FileProcessor
 {
     /**
-     * @var \App\Config\Env $env
-     */
-    private $env;
-
-    /**
      * @var \App\Service\Filesystem $filesystem
      */
     private $filesystem;
@@ -32,18 +27,15 @@ class FileProcessor
 
     /**
      * FileProcessor constructor.
-     * @param \App\Config\Env $env
      * @param \App\Service\Filesystem $filesystem
      * @param Shell $shell
      * @param DomainValidator $domainValidator
      */
     public function __construct(
-        \App\Config\Env $env,
         \App\Service\Filesystem $filesystem,
         \App\Service\Shell $shell,
         \App\Service\DomainValidator $domainValidator
     ) {
-        $this->env = $env;
         $this->filesystem = $filesystem;
         $this->shell = $shell;
         $this->domainValidator = $domainValidator;
@@ -57,18 +49,25 @@ class FileProcessor
      * @param string $applicationContainerName
      * @param string $mysqlContainer
      * @param string $phpVersion
-     * @return void
+     * @param ?string $elasticsearchVersion
+     * @param ?string $executionEnvironment
      */
     public function processDockerCompose(
         array $files,
         array $domains,
         string $applicationContainerName,
         string $mysqlContainer,
-        string $phpVersion
+        string $phpVersion,
+        ?string $elasticsearchVersion = null,
+        ?string $executionEnvironment = null
     ): void {
         $files = array_filter($files, static function ($file) {
             return preg_match('/docker-.*\.yml/', $file);
         });
+
+        if ($executionEnvironment) {
+            $dockerfile = $this->filesystem->copyDockerfile($phpVersion, $executionEnvironment);
+        }
 
         foreach ($files as $file) {
             $content = str_replace(
@@ -77,7 +76,7 @@ class FileProcessor
                     'example.com www.example.com example-2.com www.example-2.com',
                     'container_name: example.com',
                     'serverName=example.com',
-                    'example.com',
+                    'example-com',
                     'mysql57:mysql',
                     'php:version'
                 ],
@@ -92,6 +91,58 @@ class FileProcessor
                 ],
                 file_get_contents($file)
             );
+
+            // Fast implementation to support Magento 2.4.0. Plan that later compose files will become modular.
+            if ($elasticsearchVersion && strpos($file, 'docker-compose') === 0) {
+                $content .= <<<YAML
+
+                  elasticsearch:
+                    image: docker.elastic.co/elasticsearch/elasticsearch:$elasticsearchVersion
+                    environment:
+                      - network.host=0.0.0.0
+                      - http.host=0.0.0.0
+                      - transport.host=127.0.0.1
+                      - xpack.security.enabled=false
+                      - indices.query.bool.max_clause_count=10240
+                      - ES_JAVA_OPTS=-Xms1024m -Xmx1024m
+                    ulimits:
+                      memlock:
+                        soft: -1
+                        hard: -1
+                    restart: always
+                    network_mode: bridge
+                YAML;
+
+                $content = str_replace(
+                    [
+                        '#    links:',
+                        '#      - elasticsearch',
+                    ],
+                    [
+                        '    links:',
+                        '      - elasticsearch',
+                    ],
+                    $content
+                );
+            }
+
+            if ($executionEnvironment) {
+                $content = str_replace(
+                    [
+                        'image: defaultvalue/php',
+                        '#    build:',
+                        '#      context: .',
+                        '#      dockerfile: docker/Dockerfile'
+                    ],
+                    [
+                        '# image: defaultvalue/php',
+                        '    build:',
+                        '      context: .',
+                        "      dockerfile: docker/$dockerfile"
+                    ],
+                    $content
+                );
+            }
 
             // If MacOS
             // if (PHP_OS === 'Darwin') {}